src/data_disposer/requirements.txt
pandas==1.1.5
numpy==1.19.5
scikit-learn==0.24.1
scipy==1.5.4
joblib==1.0.0
cassandra-driver==3.25.0
python-dotenv